Add 'create' action to Comic Controller
The client sends a POST request to /comics/{comicId}?userId={userId}.
The controller checks if there's a user with an id equal to userId
and throws an exception if there isn't one.
The comic information is obtained from the Marvel API, using the
comicId, and is converted to a MarvelObject object.
This object is converted to a ComicDTO object that is validated using a
Validator object.
After that, the comic is saved only if:
1) There's no comic in the database with the same ISBN
2) There's a comic in the database with the same ISBN and the comicId
   is also the same (this allows associating multiple users with the
   same comic)

Other changes:
- New exception handlers
  - ConstraintViolationException: to validate the ComicDTO without
    @Valid annotations, after getting the info from Marvel API
  - MissingServletRequestParameterException: when the userId is missing
    from the request
  - InvalidUserIdException (custom exception): when there's no user
    with the provided userId
  - InvalidIsbnException (custom exception): when the ISBN from Marvel
    API and the comicId provided don't match